Disentangled squad from wave spawning system.  Broke wave spawning up into two parts: plan and spawn (to allow preview later).  Developed basic distribution system for enemy squads.

+++ core/src/com/sojourn/game/entity/Entity.java
@@ -77,7 +77,7 @@ abstract public class Entity
         return health;
     }
 
-    public float getValue()    {
+    public int getValue()    {
         // Applies additional scalars, such as if it is an elite to the base unit archetype's value
         return getValueBase();
     }
